requirements.txt
Django==1.7.6
django-auth-kerberos==1.2.3
django-braces==1.4.0
django-cors-headers==1.0.0
django-oauth-toolkit==0.7.2
djangorestframework==3.1.0
oauthlib==0.7.2
pykerberos==1.1.5
six==1.9.0
wsgiref==0.1.2
